fix: Move ErrorLogger trait to the only place it's used
